+++ src/dfxvm/cli.rs
@@ -42,14 +42,14 @@ pub struct DefaultOpts {
     version: Option<Version>,
 }
 
-/// List installed or remote available versions of dfx
+/// List installed or available versions of dfx
 #[derive(Parser)]
 pub struct ListOpts {
-    /// Specifies to list the remote available versions.
+    /// Specifies to list the available versions.
     #[arg(long)]
     pub available: bool,
 
-    /// Specifies the number of remote versions to be listed, in reverse chronological order.
+    /// Specifies the number of available versions to be listed, in reverse chronological order.
     #[arg(long, default_value_t = 10, requires("available"))]
     pub limit: usize,
 }